requirements.txt
absl-py>=2.0.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.11.18
aiosignal>=1.3.2
altair>=5.1.2
annotated-types>=0.6.0
anyio>=4.1.0
argon2-cffi>=23.1.0
argon2-cffi-bindings>=21.2.0
arrow>=1.3.0
asttokens>=2.4.1
astunparse>=1.6.3
async-lru>=2.0.4
async-timeout>=5.0.1
attrs>=23.1.0
Babel>=2.13.1
beautifulsoup4>=4.12.2
bleach>=6.1.0
blinker>=1.9.0
blis>=0.7.11
cachetools>=5.3.2
catalogue>=2.0.10
certifi>=2023.11.17
cffi>=1.16.0
charset-normalizer>=3.3.2
click>=8.1.7
cloudpathlib>=0.16.0
colorama>=0.4.6
colorlover>=0.3.0
comm>=0.2.0
confection>=0.1.4
contourpy>=1.2.0
cufflinks>=0.17.3
cycler>=0.12.1
cymem>=2.0.8
debugpy>=1.8.0
decorator>=5.1.1
defusedxml>=0.7.1
deprecation>=2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup>=1.2.0
executing>=2.0.1
fastjsonschema>=2.19.0
filetype>=1.2.0
Flask>=3.1.0
flask-cors>=5.0.1
flatbuffers>=23.5.26
fonttools>=4.45.1
fqdn>=1.5.1
frozendict>=2.4.6
frozenlist>=1.6.0
gast>=0.5.4
gensim>=4.3.2
gitdb>=4.0.11
GitPython>=3.1.40
google-ai-generativelanguage>=0.6.18
google-api-core>=2.24.2
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-pasta>=0.2.0
googleapis-common-protos>=1.70.0
gotrue>=2.12.0
grpcio>=1.72.0rc1
grpcio-status>=1.72.0rc1
groq>=0.12.0
h11>=0.16.0
h2>=4.2.0
h5py>=3.10.0
hpack>=4.1.0
html5lib>=1.1
httpcore>=1.0.9
httpx>=0.28.1
hyperframe>=6.1.0
idna>=3.6
imbalanced-learn>=0.11.0
imblearn>=0.0
importlib-metadata>=6.8.0
importlib-resources>=6.1.1
iniconfig>=2.1.0
ipykernel>=6.26.0
ipython>=8.18.0
ipywidgets>=8.1.1
isoduration>=20.11.0
itsdangerous>=2.2.0
jedi>=0.19.1
Jinja2>=3.1.2
joblib>=1.3.2
json5>=0.9.14
jsonpatch>=1.33
jsonpointer>=2.4
jsonschema>=4.20.0
jsonschema-specifications>=2023.11.1
jupyter>=1.0.0
jupyter-console>=6.6.3
jupyter-events>=0.9.0
jupyter-lsp>=2.2.1
jupyter_client>=8.6.0
jupyter_core>=5.5.0
jupyter_server>=2.10.1
jupyter_server_terminals>=0.4.4
jupyterlab>=4.0.9
jupyterlab-widgets>=3.0.9
jupyterlab_pygments>=0.3.0
jupyterlab_server>=2.25.2
keras>=2.15.0
keras-models>=0.0.7
kiwisolver>=1.4.5
lab>=8.0
langchain-core>=0.3.59
langchain-google-genai>=2.1.4
langcodes>=3.3.0
langsmith>=0.3.42
libclang>=16.0.6
lxml>=5.3.0
Markdown>=3.5.1
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
matplotlib>=3.8.2
matplotlib-inline>=0.1.6
mdurl>=0.1.2
mistune>=3.0.2
mizani>=0.9.3
ml-dtypes>=0.2.0
mnist>=0.2.2
multidict>=6.4.4
multitasking>=0.0.11
murmurhash>=1.0.10
nbclient>=0.9.0
nbconvert>=7.11.0
nbformat>=5.9.2
nest-asyncio>=1.5.8
nltk>=3.8.1
notebook>=7.0.6
notebook_shim>=0.2.3
numpy>=1.26.2
oauthlib>=3.2.2
opencv-python>=4.8.1.78
openslide-python>=1.3.1
opt-einsum>=3.3.0
orjson>=3.10.18
outcome>=1.3.0.post0
overrides>=7.4.0
packaging>=23.2
pandas>=2.1.3
pandocfilters>=1.5.0
parso>=0.8.3
pathlib>=1.0.1
patsy>=0.5.5
peewee>=3.17.8
Pillow>=10.1.0
platformdirs>=4.0.0
plotly>=5.18.0
plotnine>=0.12.4
pluggy>=1.6.0
postgrest>=1.0.1
preshed>=3.0.9
prometheus-client>=0.19.0
prompt-toolkit>=3.0.41
propcache>=0.3.1
proto-plus>=1.26.1
protobuf>=6.31.0rc2
psutil>=5.9.6
psycopg2-binary>=2.9.10
pure-eval>=0.2.2
pyarrow>=14.0.1
pyasn1>=0.5.1
pyasn1-modules>=0.3.0
pycparser>=2.21
pydantic>=2.5.2
pydantic_core>=2.14.5
pydeck>=0.8.1b0
Pygments>=2.17.2
PyJWT>=2.10.1
PyMuPDF>=1.25.5
pyparsing>=3.1.1
PySocks>=1.7.1
pytest>=8.3.5
pytest-mock>=3.14.0
python-dateutil>=2.8.2
python-docx>=1.1.2
python-dotenv>=1.0.1
python-json-logger>=2.0.7
pytz>=2023.3.post1
PyYAML>=6.0.1
pyzmq>=25.1.1
qtconsole>=5.5.1
QtPy>=2.4.1
realtime>=2.4.3
referencing>=0.31.0
regex>=2023.12.25
requests>=2.31.0
requests-oauthlib>=1.3.1
requests-toolbelt>=1.0.0
rfc3339-validator>=0.1.4
rfc3986-validator>=0.1.1
rich>=13.7.0
rpds-py>=0.13.1
rsa>=4.9
scikit-learn>=1.3.2
scipy>=1.11.4
seaborn>=0.13.1
selenium>=4.27.1
Send2Trash>=1.8.2
simplejson>=3.19.2
six>=1.16.0
smart-open>=6.4.0
smmap>=5.0.1
smote>=0.1
sniffio>=1.3.0
sortedcontainers>=2.4.0
soupsieve>=2.5
spacy>=3.7.2
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
srsly>=2.4.8
stack-data>=0.6.3
statsmodels>=0.14.1
storage3>=0.11.3
streamlit>=1.28.2
StrEnum>=0.4.15
supabase>=2.15.1
supafunc>=0.9.4
tenacity>=8.2.3
termcolor>=2.3.0
terminado>=0.18.0
thinc>=8.2.1
threadpoolctl>=3.2.0
tinycss2>=1.2.1
toml>=0.10.2
tomli>=2.0.1
toolz>=0.12.0
tornado>=6.3.3
tqdm>=4.66.1
traitlets>=5.13.0
trio>=0.23.2
trio-websocket>=0.11.1
txt2tags>=3.9
typer>=0.9.0
types-python-dateutil>=2.8.19.14
typing_extensions>=4.13.2
tzdata>=2023.3
tzlocal>=5.2
uri-template>=1.3.0
urllib3>=2.1.0
validators>=0.22.0
wasabi>=1.1.2
watchdog>=3.0.0
wcwidth>=0.2.12
weasel>=0.3.4
webcolors>=1.13
webdriver-manager>=4.0.2
webencodings>=0.5.1
websocket-client>=1.8.0
websockets>=14.2
Werkzeug>=3.1.3
widgetsnbextension>=4.0.9
wordcloud>=1.9.3
wrapt>=1.14.1
wsproto>=1.2.0
xgboost>=2.0.3
yarl>=1.20.0
yfinance>=0.2.51
zipp>=3.17.0
zstandard>=0.23.0
